Revert "Created a class file to implement TableDataReader"
This reverts commit 47121b2532a90e40c73d76bb12b04013074dc1ec.
