+++ src/main/java/org/openrefine/extensions/sparql/utils/SPARQLImportingController.java
@@ -3,6 +3,8 @@ package org.openrefine.extensions.sparql.utils;
 
 import java.io.IOException;
 import java.io.Writer;
+import java.util.ArrayList;
+import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Properties;
@@ -16,18 +18,26 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.refine.ProjectManager;
 import com.google.refine.ProjectMetadata;
 import com.google.refine.RefineServlet;
 import com.google.refine.commands.HttpUtilities;
 import com.google.refine.importers.TabularImportingParserBase;
+import com.google.refine.importers.TabularImportingParserBase.TableDataReader;
 import com.google.refine.importing.ImportingController;
 import com.google.refine.importing.ImportingJob;
 import com.google.refine.importing.ImportingManager;
 import com.google.refine.model.Project;
 import com.google.refine.util.JSONUtilities;
 import com.google.refine.util.ParsingUtilities;
+
+import okhttp3.HttpUrl;
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import okhttp3.Response;
 
 public class SPARQLImportingController implements ImportingController {
 
@@ -190,6 +200,8 @@ public class SPARQLImportingController implements ImportingController {
         String endpoint = options.get("endpoint").asText();
         String query = options.get("query").asText();
 
+        setProgress(job, "SPARQL", 0);
+
         TabularImportingParserBase.readTable(
                 project,
                 job,
@@ -197,6 +209,125 @@ public class SPARQLImportingController implements ImportingController {
                 limit,
                 options,
                 exceptions);
+        setProgress(job, "SPARQL", 100);
+
+    }
+
+    static private void setProgress(ImportingJob job, String category, int percent) {
+        job.setProgress(percent, "Reading " + category);
+    }
+
+    static protected class SPARQLQueryResultPreviewReader implements TableDataReader {
+
+        final ImportingJob job;
+        String endpoint;
+        HttpUrl urlBase;
+        JsonNode results;
+        String query;
+        private List<List<Object>> rowsOfCells = null;
+        JsonNode firstEntry;
+        List<String> columnNames = new ArrayList<String>();
+        private int batchRowStart = 0;
+        private boolean end = false;
+        private boolean usedHeaders = false;
+        private int nextRow = 0;
+        private final int batchSize;
+        List<String> jsonRows = new ArrayList<String>();
+
+        public SPARQLQueryResultPreviewReader(ImportingJob job, String endpoint, String query, int batchSize) throws IOException {
+
+            this.job = job;
+            this.endpoint = endpoint;
+            this.query = query;
+            this.batchSize = batchSize;
+            getResults();
+
+        }
+
+        public void getResults() throws IOException {
+
+            OkHttpClient client = new OkHttpClient.Builder().build();
+            urlBase = HttpUrl.parse(endpoint).newBuilder()
+                    .addQueryParameter("query", query)
+                    .addQueryParameter("format", "json").build();
+
+            Request request = new Request.Builder().url(urlBase).build();
+            Response response = client.newCall(request).execute();
+            JsonNode jsonNode = new ObjectMapper().readTree(response.body().string());
+            results = jsonNode.path("results").path("bindings");
+            resultSize = results.size();
+            firstEntry = results.get(0);
+            Iterator<String> iterator = firstEntry.fieldNames();
+            iterator.forEachRemaining(e -> columnNames.add(e));
+
+            for (int i = 0; i < resultSize; i++) {
+                JsonNode jsonObject = results.get(i);
+                Iterator<JsonNode> nodeIterator = jsonObject.elements();
+                nodeIterator.forEachRemaining(valueNode -> {
+                    if (valueNode.has("value"))
+                        jsonRows.add(valueNode.get("value").asText());
+                });
+            }
+
+        }
+
+        @Override
+        public List<Object> getNextRowOfCells() throws IOException {
+
+            if (!usedHeaders) {
+                List<Object> row = new ArrayList<Object>();
+                for (String columnName : columnNames) {
+                    row.add(columnName);
+                }
+                usedHeaders = true;
+
+                return row;
+            }
+
+            if (rowsOfCells == null || (nextRow >= batchRowStart + rowsOfCells.size() && !end)) {
+                int newBatchRowStart = batchRowStart + (rowsOfCells == null ? 0 : rowsOfCells.size());
+                rowsOfCells = getRowsOfCells(newBatchRowStart);
+                batchRowStart = newBatchRowStart;
+                setProgress(job, "SPARQL", -1);
+            }
+
+            if (rowsOfCells != null && nextRow - batchRowStart < rowsOfCells.size()) {
+                return rowsOfCells.get(nextRow++ - batchRowStart);
+            } else {
+                if (logger.isDebugEnabled()) {
+                    logger.debug("nextRow:{}, batchRowStart:{}", nextRow, batchRowStart);
+                }
+
+                return null;
+            }
+
+        }
+
+        private List<List<Object>> getRowsOfCells(int newBatchRowStart) {
+
+            List<List<Object>> rowsOfCells = new ArrayList<List<Object>>(batchSize);
+
+            if (jsonRows != null && !jsonRows.isEmpty() && jsonRows.size() > 0) {
+
+                List<Object> rowOfCells = new ArrayList<Object>(jsonRows.size());
+
+                for (int j = 0; j < jsonRows.size() && j < columnNames.size(); j++) {
+
+                    String text = jsonRows.get(j);
+                    if (text == null || text.isEmpty()) {
+                        rowOfCells.add(null);
+                    } else {
+
+                        rowOfCells.add(text);
+                    }
+
+                }
+                rowsOfCells.add(rowOfCells);
+
+            }
+            end = jsonRows.size() < batchSize + 1;
+            return rowsOfCells;
+        }
 
     }
 
@@ -272,6 +403,8 @@ public class SPARQLImportingController implements ImportingController {
         String endpoint = options.get("endpoint").asText();
         String query = options.get("query").asText();
 
+        setProgress(job, "Reading", 0);
+
         TabularImportingParserBase.readTable(
                 project,
                 job,
@@ -279,6 +412,7 @@ public class SPARQLImportingController implements ImportingController {
                 limit,
                 options,
                 exceptions);
+        setProgress(job, "Reading", 100);
 
     }
